Debug <radialGradient> test failures
- Handle edge cases of focal point overlaps
- Fix resolution of ObjectBoundingBox with <number> offsets
- Propagate stroke-opacity and fill-opacity
- Documented SVG2 breaking change, with non-overlapping circles making
  conical gradients.

+++ src/css/color.h
@@ -42,9 +42,9 @@ struct Color {
 
   RGBA rgba() const { return std::get<RGBA>(value); }
 
-  RGBA resolve(RGBA currentColor, double opacity) const {
+  RGBA resolve(RGBA currentColor, float opacity) const {
     RGBA value = isCurrentColor() ? currentColor : rgba();
-    if (opacity != 1.0) {
+    if (opacity != 1.0f) {
       value.a = static_cast<uint8_t>(value.a * opacity);
     }
     return value;
